fixed modification and commit
